Support syncing remotes via Rclone for Borg Backups:
The new borg-rclone-home-backup-to-gdrive.sh script will perform an rclone sync to whatever is configured in `config.json` for the current distro, as determined by `get-distro.sh`.

If a borg repo and or a rclone remote is not configured, the sync script will exit 0 and not generate any desktop notifications. This is to keep things clean for environments not using the sync job.

 * Update docs about config.json changes (borgRepo.chromeOS now)
 * Add Section to README on using Rclone Sync Remotes instead.
 * Add rclone to linux-common.
 * borg-home-backup.sh now calls borg-rclone-home-backup-to-gdrive.sh. This impacts what is run for the borg backup service timer.
 * borg-home-backup.sh and borg-rclone-home-backup-to-gdrive.sh are now symlinks instead of being copied to /usr/local/bin.
 * update sample-config.json.
 * Add a get-distro.sh script -- may want to refactor later to use elsewhere.