Logitech (solaar) pop mouse show all windows.
I've had this working for a while, but was waiting for the PPA to be updated again vs installing from git manually (did it locally but never commited to beleyenv). The git PPA for solaar is now being updated again and has the new device rules to allow the show all windows behavior I wanted from my pop mouse while also having the pop keyboard.

Note that this change is somewhat breaking (if anyone happens to be using the main index script). You'll want to run `sudo add-apt-repository --remove ppa:solaar-unifying/stable` before re-running the changes in `index.ubuntu.sh`