syncNotes snapshot backups
At least every 30 minutes, backups of notes will be made and retained for two months. Should make it easy to rollback in case anything goes overtly wrong with syncing with rclone and foldersync pro.

Also fixes small issues with launchd. I want it on an interval, not daily at a certain minute mark.